# web/docs/static/openapi/public.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "description": "An error the API answers with. `code` is stable, and what an application should act on; `error` is its English sentence; `params` fills in the sentence's {placeholders}.",
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "type": "object"
      },
...
      "SessionInfo": {
        "description": "SessionInfo is one browser a user is signed in on.",
        "properties": {
          "current": {
            "description": "`current` marks the session the request itself carries.",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "signed_in_at": {
            "format": "date-time",
            "type": "string"
          },
          "user_agent": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ip",
          "user_agent",
          "signed_in_at",
          "expires_at",
          "current"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "accessToken": {
        "bearerFormat": "JWT",
        "description": "An access token this server issued.",
        "scheme": "bearer",
        "type": "http"
      },
      "accountToken": {
        "bearerFormat": "JWT",
        "description": "A user's access token for the account API (audience urn:loginer:account-api), with account.read to read and account.write to change.",
        "scheme": "bearer",
        "type": "http"
      },
      "clientSecretBasic": {
        "description": "An application's client_id and client_secret, form-encoded and then base64-encoded (RFC 6749 section 2.3.1). The same credentials may be sent in the form instead.",
        "scheme": "basic",
        "type": "http"
      },
      "userSession": {
        "description": "A user's session at the provider, set by signing in on the sign-in pages.",
        "in": "cookie",
        "name": "loginer_user_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The OAuth 2.0 and OpenID Connect provider applications sign their users in with, and the account API the sign-in pages call.",
    "title": "Loginer Public API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/account/sessions": {
      "get": {
        "operationId": "accountSessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "sessions": {
                      "items": {
                        "$ref": "#/components/schemas/SessionInfo"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "sessions"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "account_token_refused": {
                    "summary": "This access token is not accepted by the account API: it has expired, is for another API, or its user can no longer sign in.",
                    "value": {
                      "code": "account_token_refused",
                      "error": "This access token is not accepted by the account API: it has expired, is for another API, or its user can no longer sign in."
                    }
                  },
                  "not_signed_in": {
                    "summary": "You are not signed in.",
                    "value": {
                      "code": "not_signed_in",
                      "error": "You are not signed in."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized: `account_token_refused`, `not_signed_in`"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "account_scope_missing": {
                    "summary": "This access token needs the {scope} scope.",
                    "value": {
                      "code": "account_scope_missing",
                      "error": "This access token needs the {scope} scope."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden: `account_scope_missing`"
          }
        },
        "security": [
          {
            "userSession": []
          },
          {
            "accountToken": [
              "account.read"
            ]
          }
        ],
        "summary": "Lists where the signed-in user is signed in",
        "tags": [
          "Account"
        ]
      }
    },
    "/api/v1/account/sessions/{id}": {
      "delete": {
        "operationId": "accountEndSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "account_token_refused": {
                    "summary": "This access token is not accepted by the account API: it has expired, is for another API, or its user can no longer sign in.",
                    "value": {
                      "code": "account_token_refused",
                      "error": "This access token is not accepted by the account API: it has expired, is for another API, or its user can no longer sign in."
                    }
                  },
                  "not_signed_in": {
                    "summary": "You are not signed in.",
                    "value": {
                      "code": "not_signed_in",
                      "error": "You are not signed in."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized: `account_token_refused`, `not_signed_in`"
          },
          "403": {
            "content": {
              "application/json": {
                "examples": {
                  "account_scope_missing": {
                    "summary": "This access token needs the {scope} scope.",
                    "value": {
                      "code": "account_scope_missing",
                      "error": "This access token needs the {scope} scope."
                    }
                  },
                  "cross_origin": {
                    "summary": "Requests from {origin} are not allowed.",
                    "value": {
                      "code": "cross_origin",
                      "error": "Requests from {origin} are not allowed."
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden: `account_scope_missing`, `cross_origin`"
          }
        },
        "security": [
          {
            "userSession": []
          },
          {
            "accountToken": [
              "account.write"
            ]
          }
        ],
        "summary": "Signs one of the user's other browsers out",
        "tags": [
          "Account"
        ]
      }
    },
...
  },
...
}